fixed duties description box word wrap and validation
